server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
argcomplete==1.10.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.8.2
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
compressed-rtf==1.0.6
dataclasses-json==0.6.7
dnspython==2.6.1
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
faiss-cpu==1.8.0.post1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
huggingface-hub==0.24.2
idna==3.7
IMAPClient==2.1.0
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.23
langchain-google-genai==1.0.8
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langsmith==0.1.93
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
olefile==0.47
orjson==3.10.6
packaging==24.1
pdfminer.six==20191110
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-pptx==0.6.23
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.31
sympy==1.13.1
tenacity==8.5.0
textract==1.6.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
transformers==4.43.2
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
xlrd==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.19.2
